Make fishhook work reliably on arm7 devices.
To make fishhook work on arm7 devices we need to get the fully resolved symbol.  The trick is to call the function before calling rebind_symbols.  This ensures that the symbol is fully resolved in the lazy lookup table and doesn't point to a stub.  We then grab the address of the fully resolved symbol before replacing it with our replacement function.  We need to know which module we can guarantee has the symbol fully resolved, so that's passed as an input parameter.

Note: The input rebinding struct has changed.